reorganize integrationtest packages & use descriptive/actionable test names
